chore: moved JoinRoom to pages/joinRoom AND applied shared-type in server.ts
